+++ web/wiki/src/hooks/useUpdateResource.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import { useMutation } from '@tanstack/react-query'
 import { useForm } from 'react-hook-form'
 import { updateResourceFetcher } from '../helpers/fetchers'
@@ -6,21 +5,16 @@ import { reloadPage } from '../utils/navigation'
 import type { TResourceForm } from '../components/organisms/ResourceForm'
 
 export const useUpdateResource = () => {
-  const [responseError, setResponseError] = useState('')
   const { reset } = useForm<TResourceForm>()
   const updateResource = useMutation(updateResourceFetcher, {
     onSuccess: () => {
       reset()
       reloadPage()
     },
-    onError: (error: Error) => {
-      setResponseError(error.message)
-    },
   })
 
   return {
     updateResource,
-    responseError,
     isLoading: updateResource.isLoading,
     isSuccess: updateResource.isSuccess,
   }