+++ web/wiki/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.20.2] - 2024-06-11
+## [0.20.2] - 2024-06-16
 
 ### Fixed
 
-- Manage errors in ResourceForm for create and update resource
-- Topic selector error disappearance
+- Fixed manage errors in ResourceForm for create and update resource
+- Fixed topic selector error disappearance
 
 ## [0.20.1] - 2024-06-06